add error checking if inventory or item not found

+++ flaskApp/views.py
@@ -1,5 +1,5 @@
 # import render_template for the static index page
-from flask import render_template, jsonify
+from flask import render_template, jsonify, abort
 
 # import the app, the database, and the api manager
 from flaskApp import app, db, manager
@@ -25,6 +25,21 @@ def list_inventories():
     return jsonify(inventories = inventories)
 
 
+
+# list items in named inventory
+@app.route('/api/inventory/<name>', methods=['GET'])
+def get_items(name):
+    # query for the specific inveontory
+    qryresult = Container.query.filter_by(name = name).first()
+    if qryresult == None:
+        abort(404)
+    # get and serialize items from inveontory
+    items = [i.serialize() for i in qryresult.items]
+    # return as json
+    return jsonify(items = items)
+
+
+
 # modify/add inventory
 
 # delete inventory
@@ -36,25 +51,24 @@ def delete_inventory(name):
         db.session.commit()
         # return name of deleted container
         return jsonify(deleted = {'name' : name})
+    else:
+        abort(404)
 
 
-# list items in named inventory
-@app.route('/api/inventory/<name>', methods=['GET'])
-def get_items(name):
-    # query for the specific inveontory
-    qryresult = Container.query.filter_by(name = name).first()
-    # get and serialize items from inveontory
-    items = [i.serialize() for i in qryresult.items]
-    # return as json
-    return jsonify(items = items)
 
 # get specific item
 @app.route('/api/inventory/<name>/<int:item_id>', methods=['GET'])
 def get_item(name, item_id):
-    # get containerId from name
-    cid = Container.query.filter_by(name = name).first().id
+    # get container if exists
+    c = Container.query.filter_by(name = name).first()
+    if c == None:
+        abort(404)
+    # get containerId from result
+    cid = c.id
     # get items within containerId
     qryresult = Item.query.filter_by(id = item_id, containerId = cid).first()
+    if qryresult == None:
+        abort(404)
     # return the one item serialized
     return jsonify(qryresult.serialize())
 
@@ -63,8 +77,12 @@ def get_item(name, item_id):
 # delete item
 @app.route('/api/inventory/<name>/<int:item_id>', methods=['DELETE'])
 def delete_item(name, item_id):
-    # get containerId from name
-    cid = Container.query.filter_by(name = name).first().id
+    # get container if exists
+    c = Container.query.filter_by(name = name).first()
+    if c == None:
+        abort(404)
+    # get containerId from result
+    cid = c.id
     # check to make sure 1 and only one record
     if Item.query.filter_by(id = item_id, containerId = cid).delete() == 1:
         # if one record commit transaction
@@ -72,6 +90,8 @@ def delete_item(name, item_id):
         # return name of deleted container
         return jsonify(deleted = { 'item_id' : item_id,
                 'inventory_name' : name})
+    else:
+        abort(404)
